CSRF MW Restructuring
